third_party/WebKit/LayoutTests/inspector/console/console-format-table-expected.txt

CONSOLE MESSAGE: line 10: null
CONSOLE MESSAGE: line 12: Array of arrays
CONSOLE MESSAGE: line 13: 1,2,3,4,5,6
CONSOLE MESSAGE: line 15: Large array of arrays
CONSOLE MESSAGE: line 16: 1,2,3,4,5,6,1,2,3,4,5,6,1,2,3,4,5,6,1,2,3,4,5,6
CONSOLE MESSAGE: line 18: Array or array and object
CONSOLE MESSAGE: line 19: 1,2,3,[object Object]
CONSOLE MESSAGE: line 21: Object table
CONSOLE MESSAGE: line 22: [object Object]
CONSOLE MESSAGE: line 24: Null as columns
CONSOLE MESSAGE: line 25: 1,2,3,4,5,6
CONSOLE MESSAGE: line 27: Digit as columns
CONSOLE MESSAGE: line 28: 1,2,3,4,5,6
CONSOLE MESSAGE: line 30: String as columns
CONSOLE MESSAGE: line 31: 1,2,3,4,5,6
CONSOLE MESSAGE: line 33: Random string as columns
CONSOLE MESSAGE: line 34: 1,2,3,4,5,6
CONSOLE MESSAGE: line 36: Array of strings as columns
CONSOLE MESSAGE: line 37: [object Object],[object Object]
CONSOLE MESSAGE: line 39: Good and bad column names
CONSOLE MESSAGE: line 40: [object Object],[object Object]
CONSOLE MESSAGE: line 42: Missing column name
CONSOLE MESSAGE: line 43: [object Object],[object Object]
CONSOLE MESSAGE: line 10: null
CONSOLE MESSAGE: line 12: Array of arrays
CONSOLE MESSAGE: line 13: 1,2,3,4,5,6
CONSOLE MESSAGE: line 15: Large array of arrays
CONSOLE MESSAGE: line 16: 1,2,3,4,5,6,1,2,3,4,5,6,1,2,3,4,5,6,1,2,3,4,5,6
CONSOLE MESSAGE: line 18: Array or array and object
CONSOLE MESSAGE: line 19: 1,2,3,[object Object]
CONSOLE MESSAGE: line 21: Object table
CONSOLE MESSAGE: line 22: [object Object]
CONSOLE MESSAGE: line 24: Null as columns
CONSOLE MESSAGE: line 25: 1,2,3,4,5,6
CONSOLE MESSAGE: line 27: Digit as columns
CONSOLE MESSAGE: line 28: 1,2,3,4,5,6
CONSOLE MESSAGE: line 30: String as columns
CONSOLE MESSAGE: line 31: 1,2,3,4,5,6
CONSOLE MESSAGE: line 33: Random string as columns
CONSOLE MESSAGE: line 34: 1,2,3,4,5,6
CONSOLE MESSAGE: line 36: Array of strings as columns
CONSOLE MESSAGE: line 37: [object Object],[object Object]
CONSOLE MESSAGE: line 39: Good and bad column names
CONSOLE MESSAGE: line 40: [object Object],[object Object]
CONSOLE MESSAGE: line 42: Missing column name
CONSOLE MESSAGE: line 43: [object Object],[object Object]
Tests console.table.

 console-format-table.html:10
Array of arrays console-format-table.html:12
 console-format-table.html:13
Large array of arrays console-format-table.html:15
 console-format-table.html:16
Array or array and object console-format-table.html:18
 console-format-table.html:19
Object table console-format-table.html:21
 console-format-table.html:22
Null as columns console-format-table.html:24
 console-format-table.html:25
Digit as columns console-format-table.html:27
 console-format-table.html:28
String as columns console-format-table.html:30
 console-format-table.html:31
Random string as columns console-format-table.html:33
 console-format-table.html:34
Array of strings as columns console-format-table.html:36
 console-format-table.html:37
Good and bad column names console-format-table.html:39
 console-format-table.html:40
Missing column name console-format-table.html:42
 console-format-table.html:43
 console-format-table.html:10
Array of arrays console-format-table.html:12
HEADER (index) | 0 | 1 | 2 | 
ROW 1 | 2 | 3 | 
ROW 4 | 5 | 6 | 
Large array of arrays console-format-table.html:15
HEADER (index) | 0 | 1 | 2 | 
ROW 1 | 2 | 3 | 
ROW 4 | 5 | 6 | 
ROW 1 | 2 | 3 | 
ROW 4 | 5 | 6 | 
ROW 1 | 2 | 3 | 
ROW 4 | 5 | 6 | 
ROW 1 | 2 | 3 | 
ROW 4 | 5 | 6 | 
Array or array and object console-format-table.html:18
HEADER (index) | 0 | 1 | 2 | a | b | c | 
ROW 1 | 2 | 3 | 
ROW 1 | 2 | 3 | 
Object table console-format-table.html:21
HEADER (index) | a | b | 
ROW 1 | 2 | 
ROW 3 | 4 | 
Null as columns console-format-table.html:24
HEADER (index) | 0 | 1 | 2 | 
ROW 1 | 2 | 3 | 
ROW 4 | 5 | 6 | 
Digit as columns console-format-table.html:27
HEADER (index) | 0 | 1 | 2 | 
ROW 1 | 2 | 3 | 
ROW 4 | 5 | 6 | 
String as columns console-format-table.html:30
HEADER (index) | 0 | 
ROW 1 | 
ROW 4 | 
Random string as columns console-format-table.html:33
HEADER (index) | 
Array of strings as columns console-format-table.html:36
HEADER (index) | a | b | 
ROW 1 | 2 | 
ROW "foo" | "bar" | 
Good and bad column names console-format-table.html:39
HEADER (index) | a | b | 
ROW 1 | 2 | 
ROW "foo" | 
Missing column name console-format-table.html:42
HEADER (index) | 
